Implement warning diagnostic flags for RGBFIX
