+++ Makefile
@@ -97,7 +97,9 @@ src/link/main.o: src/link/script.hpp
 
 rgbfix_obj := \
 	${common_obj} \
-	src/fix/main.o
+	src/fix/main.o \
+	src/fix/mbc.o \
+	src/fix/warning.o
 
 rgbgfx_obj := \
 	${common_obj} \
