+++ contrib/bash_compl/_rgbgfx.bash
@@ -7,37 +7,38 @@ _rgbgfx_completions() {
 	# Empty long opt = it doesn't exit
 	# See the `state` variable below for info about `state_after`
 	declare -A opts=(
-		[V]="version:normal"
 		[h]="help:normal"
-		[C]="color-curve:normal"
-		[m]="mirror-tiles:normal"
-		[O]="group-outputs:normal"
-		[u]="unique-tiles:normal"
-		[v]="verbose:normal"
-		[X]="mirror-x:normal"
-		[Y]="mirror-y:normal"
-		[Z]="columns:normal"
-		[a]="attr-map:glob-*.attrmap"
+		[V]="version:normal"
+		[W]="warning:warning"
+		[w]=":normal"
 		[A]="auto-attr-map:normal"
+		[a]="attr-map:glob-*.attrmap"
 		[B]="background-color:unk"
 		[b]="base-tiles:unk"
+		[C]="color-curve:normal"
 		[c]="colors:unk"
 		[d]="depth:unk"
 		[i]="input-tileset:glob-*.2bpp"
 		[L]="slice:unk"
+		[m]="mirror-tiles:normal"
 		[N]="nb-tiles:unk"
 		[n]="nb-palettes:unk"
+		[O]="group-outputs:normal"
 		[o]="output:glob-*.2bpp"
-		[p]="palette:glob-*.pal"
 		[P]="auto-palette:normal"
-		[q]="palette-map:glob-*.palmap"
+		[p]="palette:glob-*.pal"
 		[Q]="auto-palette-map:normal"
+		[q]="palette-map:glob-*.palmap"
 		[r]="reverse:unk"
 		[s]="palette-size:unk"
-		[t]="tilemap:glob-*.tilemap"
 		[T]="auto-tilemap:normal"
-		[W]="warning:warning"
+		[t]="tilemap:glob-*.tilemap"
+		[u]="unique-tiles:normal"
+		[v]="verbose:normal"
+		[X]="mirror-x:normal"
 		[x]="trim-end:unk"
+		[Y]="mirror-y:normal"
+		[Z]="columns:normal"
 	)
 	# Parse command-line up to current word
 	local opt_ena=true
